fix: update ejsRender

+++ packages/generate-material/src/ejsRenderDir.ts
@@ -2,12 +2,12 @@ import * as path from 'path';
 import * as glob from 'glob';
 import * as ejs from 'ejs';
 import * as fse from 'fs-extra';
-import { ITemplateOptions } from './type';
+import { ITemplateOptions } from './index';
 
 export default async function (dir: string, options: ITemplateOptions): Promise<void> {
   return new Promise((resolve, reject) => {
     glob(
-      '**/*.ejs',
+      '**/*.?(_)ejs',
       {
         cwd: dir,
         nodir: true,
@@ -33,56 +33,29 @@ export default async function (dir: string, options: ITemplateOptions): Promise<
           });
       }
     );
-    glob(
-      '**/*._ejs',
-      {
-        cwd: dir,
-        nodir: true,
-        dot: true,
-        ignore: ['node_modules/**'],
-      },
-      (err, files) => {
-        if (err) {
-          return reject(err);
-        }
-
-        Promise.all(
-          files.map((file) => {
-            return renderTemplateFile(file);
-          })
-        )
-          .then(() => {
-            resolve();
-          })
-          .catch((err) => {
-            reject(err);
-          });
-      }
-    );
   });
 }
 
 function renderFile(filepath: string, options: any): Promise<string> {
   return new Promise((resolve, reject) => {
-    ejs.renderFile(filepath, options, (err, result) => {
-      if (err) {
-        return reject(err);
+    if (filepath.endsWith('_ejs')) {
+      const templateFilePath = filepath.replace('_ejs', 'ejs');
+      try {
+        fse.renameSync(filepath, templateFilePath);
+      } catch (err) {
+        reject(err);
       }
-
-      fse.removeSync(filepath);
-      fse.writeFileSync(filepath.replace(/\.ejs$/, ''), result);
       resolve();
-    });
-  });
-}
-function renderTemplateFile(filepath: string): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const templateFilePath = filepath.replace('_ejs', 'ejs');
-    try {
-      fse.rename(filepath, templateFilePath);
-    } catch (err) {
-      return reject(err);
+    } else {
+      ejs.renderFile(filepath, options, (err, result) => {
+        if (err) {
+          return reject(err);
+        }
+
+        fse.removeSync(filepath);
+        fse.writeFileSync(filepath.replace(/\.ejs$/, ''), result);
+        resolve();
+      });
     }
-    resolve();
   });
 }
